django-backend/requirements.txt
bson==0.5.8
certifi==2020.11.8
chardet==3.0.4
dataclasses==0.6
defusedxml==0.6.0
Django==2.2.13
django-allauth==0.39.1
django-cors-headers==3.5.0
django-rest-auth==0.9.3
djangorestframework==3.9.1
djongo==1.3.0
idna==2.10
numpy==1.19.4
oauthlib==3.1.0
pandas==1.1.4
psycopg2-binary==2.8.6
pymongo==3.11.1
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2020.4
requests==2.25.0
requests-oauthlib==1.3.0
rest-meets-djongo==0.0.13
six==1.15.0
sqlparse==0.2.4
urllib3==1.26.2
